♻️ move LogsEventDomainContext type to own file to fix ts module isolation compatibility
