+++ packages/logs/src/domain/lifeCycle.ts
@@ -1,7 +1,8 @@
 import { AbstractLifeCycle } from '@datadog/browser-core'
-import type { Context, ErrorSource } from '@datadog/browser-core'
+import type { Context } from '@datadog/browser-core'
 import type { LogsEvent } from '../logsEvent.types'
 import type { CommonContext, RawLogsEvent } from '../rawLogsEvent.types'
+import type { LogsEventDomainContext } from '../domainContext.types'
 
 export const enum LifeCycleEventType {
   RAW_LOG_COLLECTED,
@@ -15,14 +16,6 @@ interface LifeCycleEventMap {
 
 export const LifeCycle = AbstractLifeCycle<LifeCycleEventMap>
 export type LifeCycle = AbstractLifeCycle<LifeCycleEventMap>
-
-export type LogsEventDomainContext<T extends ErrorSource> = T extends typeof ErrorSource.NETWORK
-  ? NetworkLogsEventDomainContext
-  : never
-
-type NetworkLogsEventDomainContext = {
-  isAborted: boolean
-}
 
 export interface RawLogsEventCollectedData<E extends RawLogsEvent = RawLogsEvent> {
   rawLogsEvent: E
